refactor(core): share one serve-path warn limiter across crates
The `WarnLimiter` bounding serve-path failure warns to a burst of 10 per
30-second window lived in `musefs-fuse` and was reachable only through
`reply_errno`, so the warns emitted from inside synthesis — on the same
hot path — bypassed it entirely:

- `reader.rs`: an invalid Vorbis key dropped from FLAC/Ogg synthesis,
  once per `HeaderCache` miss. The cache is byte-budgeted, so on a
  library large enough to evict, the same track re-builds and re-warns
  forever — a walk with one bad key per album logs a line per album per
  pass.
- `mapping.rs`: art over `MAX_ART_BYTES`, and a `DbArtSource::read_window`
  failure, which fires per art *window* and so turns one failing blob
  into many lines for one file.

Move the limiter into `musefs-core::warn_limit` (no FUSE dependencies —
an atomic window counter plus `WarnDecision`), route those four sites
through it, and have `reply_errno` and the `EAGAIN` load-shed warn call
into it rather than owning it. One process-wide budget is the right
semantic: the operator's concern is total serve-path log volume, not
per-crate volume. Behavior is otherwise unchanged — same burst, window,
debug fallback, and suppressed-count suffix — and the limiter's tests
move with it.

Closes #650
